+++ src/components/Insight/InsightCardInternals.js
@@ -5,9 +5,9 @@ import { Icon } from '@santiment-network/ui'
 import InsightTags from './InsightTags'
 import ProfileInfo from './ProfileInfo'
 import MultilineText from '../MultilineText/MultilineText'
-import { getSEOLinkFromIdAndTitle } from '../../pages/Insights/utils'
 import LikeBtn from '../Like/LikeBtn'
 import { getDateFormats } from '../../utils/dates'
+import { getSEOLinkFromIdAndTitle } from './utils'
 import styles from './InsightCard.module.scss'
 
 const AWAITING_APPROVAL_STATE = 'awaiting_approval'
