+++ src/components/Insight/InsightsFeatured.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { graphql } from 'react-apollo'
 import InsightCardSmall from './InsightCardSmall'
 import { FEATURED_INSIGHTS_QUERY } from './../../queries/InsightsGQL'
-import { creationDateSort } from '../../pages/Insights/utils'
+import { creationDateSort } from './utils'
 
 const InsightsTrends = ({ data: { insights = [] }, ...props }) => {
   return insights
