+++ src/components/UnAuth/UnAuth.module.scss
@@ -1,4 +1,4 @@
-@import '~@santiment-network/ui/mixins';
+@import '../../../node_modules/@santiment-network/ui/mixins';
 
 .title {
   @include text('h4');
