+++ src/pages/SonarFeed/SonarFeedPage.js
@@ -5,7 +5,7 @@ import { Link, Route, Switch, Redirect } from 'react-router-dom'
 import Tabs from '@santiment-network/ui/Tabs'
 import Loadable from 'react-loadable'
 import PageLoader from '../../components/Loader/PageLoader'
-import UnAuth from '../Insights/UnAuth'
+import UnAuth from '../../components/UnAuth/UnAuth'
 import MobileHeader from '../../components/MobileHeader/MobileHeader'
 import { isTelegramConnected } from '../../pages/UserSelectors'
 import SonarFeedHeader from './SonarFeedActions/SonarFeedHeader'
