+++ src/pages/Trends/connectedWordsEpic.js
@@ -11,8 +11,10 @@ import { ALL_INSIGHTS_BY_TAG_QUERY } from '../../queries/InsightsGQL'
 import { binarySearch } from './utils'
 import { ONE_DAY_IN_MS } from '../../utils/dates'
 import { simpleSortStrings } from '../../utils/sortMethods'
-import { getInsightTrendTagByDate } from '../../components/Insight/utils'
-import { creationDateSort } from '../Insights/utils'
+import {
+  creationDateSort,
+  getInsightTrendTagByDate
+} from '../../components/Insight/utils'
 
 const tickerCheckClb = (target, { ticker }) => target === ticker
 const tickerMoveClb = (target, { ticker }) => target < ticker
